requirements.txt
Flask==3.0.3
mysql-connector-python==8.3.0
PyYAML==6.0.1
